Expand run script to run single assets (e.g. publish_*)
